+++ packages/components/docs/LoaderDocs.js
@@ -4,13 +4,26 @@ import { Loader, Checkbox } from '../src';
 export default class LoaderDocs extends Component {
   constructor(props) {
     super(props);
-    this.state = { small: false };
+    this.state = { small: false, hasClassNames: false };
   }
 
   createStateLink(name) {
     return value => this.setState({ [name]: value });
   }
+
   render() {
+    let docsCode = `<Loader
+  small={${this.state.small}}
+/>`;
+
+    if (this.state.hasClassNames) {
+      docsCode = docsCode.replace('/>', `  classNames={{
+    loader: 'loader_33HEu6aS3s',
+    loader-spinner: 'loader-spinner_z0k4VBxn-L',
+  }}
+/>`);
+    }
+
     return (
       <section className="section">
         <div className="row">
@@ -19,14 +32,16 @@ export default class LoaderDocs extends Component {
             <p>Like a record baby</p>
           </div>
           <div className="col-md-6 p-b-2 text-xs-center">
-            <Loader small={this.state.small} />
+            <Loader
+              small={this.state.small}
+            />
           </div>
         </div>
         <div className="row">
           <div className="col-md-6">
             {/* eslint-disable react/jsx-indent */}
             <pre className="tw-docs-code">
-              {`<Loader small={${this.state.small}} />`}
+              {docsCode}
             </pre>
             {/* eslint-enable react/jsx-indent */}
           </div>
@@ -37,6 +52,14 @@ export default class LoaderDocs extends Component {
               onChange={this.createStateLink('small')}
             />
           </div>
+          <div className="col-md-6">
+            <Checkbox
+              label="Scoped className?"
+              checked={this.state.hasClassNames}
+              onChange={this.createStateLink('hasClassNames')}
+            />
+          </div>
+          <div className="col-md-6" />
         </div>
       </section>
     );